feat(SalesChannel): Add Sales Channel and Shipping Profile
Goal is to add the default Sales Channel and Shipping Profile which are currently causing errors in the admin due to the
`default_sales_channel` being `null`.

Both Sales Channel and Shipping Profile will be tied to a `store_id` so that other tenants don't see incorrect Sales Channel /
Shipping Profile.

Goal is for the `store_id` to be required, but since `medusa/src/loaders/default.ts` has `createDefault` it makes this difficult.